folder restructure / file renamings
