+++ frontend/src/components/scheduleEntry/DialogActivityCreate.vue
@@ -20,8 +20,8 @@
 </template>
 
 <script>
-import DialogForm from './DialogForm.vue'
-import DialogBase from './DialogBase.vue'
+import DialogForm from '@/components/dialog/DialogForm.vue'
+import DialogBase from '@/components/dialog/DialogBase.vue'
 import DialogActivityForm from './DialogActivityForm.vue'
 
 export default {
