+++ frontend/src/components/scheduleEntry/DialogActivityEdit.vue
@@ -23,8 +23,8 @@
 </template>
 
 <script>
-import DialogForm from './DialogForm.vue'
-import DialogBase from './DialogBase.vue'
+import DialogForm from '@/components/dialog/DialogForm.vue'
+import DialogBase from '@/components/dialog/DialogBase.vue'
 import DialogActivityForm from './DialogActivityForm.vue'
 import { scheduleEntryRoute } from '@/router.js'
