+++ frontend/src/components/scheduleEntry/FormScheduleEntryItem.vue
